+++ client-react/src/pages/app/functions/create/DetailsPivot.tsx
@@ -1,17 +1,14 @@
 import React from 'react';
 import { FunctionTemplate } from '../../../../models/functions/function-template';
 import { DefaultButton } from 'office-ui-fabric-react';
-import FunctionsService from '../../../../ApiHelpers/FunctionsService';
 import { useTranslation } from 'react-i18next';
 import { Formik, FormikProps } from 'formik';
 import { BindingConfigMetadata } from '../../../../models/functions/bindings-config';
-import { BindingInfo } from '../../../../models/functions/function-binding';
 import { CreateFunctionFormBuilder, CreateFunctionFormValues } from '../common/CreateFunctionFormBuilder';
 import { FunctionInfo } from '../../../../models/functions/function-info';
 import { ArmObj } from '../../../../models/arm-obj';
-import { FunctionConfig } from '../../../../models/functions/function-config';
-import { BindingEditorFormValues } from '../common/BindingFormBuilder';
 import { paddingStyle } from './FunctionCreate.styles';
+import { getTriggerBinding, getRequiredBindingMetadata, onCreateFunctionClicked } from './FunctionCreate.data';
 
 interface DetailsPivotProps {
   functionsInfo: ArmObj<FunctionInfo>[];
@@ -35,7 +32,7 @@ const DetailsPivot: React.FC<DetailsPivotProps> = props => {
       triggerBinding,
       requiredBindingMetadata,
       functionsInfo,
-      selectedFunctionTemplate.metadata.defaultFunctionName || 'newFunction',
+      selectedFunctionTemplate.metadata.defaultFunctionName || 'NewFunction',
       t
     );
     const initialFormValues = builder.getInitialFormValues();
@@ -50,7 +47,7 @@ const DetailsPivot: React.FC<DetailsPivotProps> = props => {
               <form>
                 <div style={paddingStyle}>
                   {builder.getFields(formProps, false)}
-                  <DefaultButton onClick={formProps.submitForm}>Create Function</DefaultButton>
+                  <DefaultButton onClick={formProps.submitForm}>{t('functionCreate_createFunction')}</DefaultButton>
                 </div>
               </form>
             );
@@ -60,64 +57,6 @@ const DetailsPivot: React.FC<DetailsPivotProps> = props => {
     );
   }
   return <></>;
-};
-
-const onCreateFunctionClicked = (
-  resourceId: string,
-  functionTemplate: FunctionTemplate,
-  triggerBinding: BindingInfo,
-  formValues: CreateFunctionFormValues
-) => {
-  const config = buildFunctionConfig(functionTemplate.function.bindings, triggerBinding, formValues);
-  FunctionsService.createFunction(resourceId, formValues.functionName, functionTemplate.files, config);
-};
-
-const getTriggerBinding = (functionTemplate: FunctionTemplate): BindingInfo => {
-  return functionTemplate.function.bindings.find(binding => binding.type.toLowerCase().includes('trigger')) as BindingInfo;
-};
-
-// Not all bindings are required for function creation
-// Only display bindings that are list in the funciton template 'userPrompt'
-const getRequiredBindingMetadata = (
-  triggerBinding: BindingInfo,
-  bindingsConfigMetatdata: BindingConfigMetadata[],
-  userPrompt: string[]
-): BindingConfigMetadata => {
-  const currentBindingMetadata = bindingsConfigMetatdata.find(b => b.type === triggerBinding.type) as BindingConfigMetadata;
-  const requiredBindings = currentBindingMetadata;
-  requiredBindings.settings = currentBindingMetadata.settings.filter(setting => {
-    return userPrompt.find(prompt => prompt === setting.name);
-  });
-  return requiredBindings;
-};
-
-const buildFunctionConfig = (
-  defaultBindingInfo: BindingInfo[],
-  triggerBinding: BindingInfo,
-  formValues: BindingEditorFormValues
-): FunctionConfig => {
-  const resultConfig: FunctionConfig = {
-    bindings: [],
-  };
-
-  defaultBindingInfo.forEach(bindingInfo => {
-    // Only look at form values for the trigger Binding
-    // Else, (when not the trigger Binding) directly copy the Binding
-    if (bindingInfo === triggerBinding) {
-      const bindingInfoCopy = bindingInfo;
-      // Update binding values that exist in the form
-      for (const key in bindingInfo) {
-        if (formValues.hasOwnProperty(key)) {
-          bindingInfoCopy[key] = formValues[key];
-        }
-      }
-      resultConfig.bindings.push(bindingInfoCopy);
-    } else {
-      resultConfig.bindings.push(bindingInfo);
-    }
-  });
-
-  return resultConfig;
 };
 
 export default DetailsPivot;
