+++ client/src/app/shared/models/portal-resources.ts
@@ -1648,4 +1648,5 @@ export class PortalResources {
   public static functionCreate_templates = 'functionCreate_templates';
   public static functionCreate_details = 'functionCreate_details';
   public static functionCreate_searchByTemplateName = 'functionCreate_searchByTemplateName';
+  public static functionCreate_createFunction = 'functionCreate_createFunction';
 }
